ssh_client: glibc-compat: split out of webports
We can use the new Python framework to build glibc-compat ourselves.
The goal is to pull the few packages webports provides out so we can
drop webports entirely.

We have to tweak the naclsdk build to sync the internal pnacl headers
over to the sysroot.  The webports code used to do this, and if it's
skipped, glibc-compat & openssh fail to build.

Note that the glibc-compat package has always been a hand curated set
of files that the nacl/webports people found helped ease porting since
the C library the pnacl toolchain included (newlib iirc) wasn't as
feature complete as the nacl toolchain (glibc).  It has many POSIX
shims that we don't actually need, as well as a full resolver library.
Even though it's called "glibc-compat", the files didn't come from
glibc, nor did they really implement glibc-specific behavior.  It was
named that way simply because the nacl toolchain used glibc, and the
pnacl toolchain used newlib, and thus a compat layer became "glibc".

This isn't a straight copy of the existing webports code we're using.
That was about 680k of source files.  This is a stripped down set of
files (about 60k) of just what openssl/openssh seems to need.  In my
testing, things seem to work OK, but we'll have to do some testing in
the field and see what users think.  If need be, we can dump all the
files back in to regain stability.  Considering the source size, I'm
hoping to avoid that.

Now that we've switched OpenSSH over to using ldns, turns out we no
longer need any of the resolver logic, so we can dump all of that.

We'll probably need to rework this package for WASM, but we'll cross
that bridge once we have a build/port working.

Reviewed-on: https://chromium-review.googlesource.com/c/1336708
Tested-by: Mike Frysinger <vapier@chromium.org>

+++ ssh_client/src/syscalls.cc
@@ -219,8 +219,8 @@ int fcntl(int fd, int cmd, ...) {
   return ret;
 }
 
-int ioctl(int fd, int request, ...) {
-  LOG("ioctl: %d %d\n", fd, request);
+int ioctl(int fd, unsigned long request, ...) {
+  LOG("ioctl: %d %lu\n", fd, request);
   va_list ap;
   va_start(ap, request);
   int ret = FileSystem::GetFileSystem()->ioctl(fd, request, ap);
